@@ -13,7 +13,6 @@ use Laminas\Stdlib\ArrayUtils;
 use Traversable;
 
 use function class_exists;
-use function get_class;
 use function gettype;
 use function is_array;
 use function is_object;
@@ -82,7 +81,7 @@ class ServiceListener implements ServiceListenerInterface
         } else {
             throw new Exception\RuntimeException(sprintf(
                 'Invalid service manager provided, expected ServiceManager or string, %s provided',
-                is_object($serviceManager) ? get_class($serviceManager) : gettype($serviceManager)
+                is_object($serviceManager) ? $serviceManager::class : gettype($serviceManager)
             ));
         }
 
@@ -243,7 +242,7 @@ class ServiceListener implements ServiceListenerInterface
         if (! $config instanceof ServiceConfigInterface) {
             throw new Exception\RuntimeException(sprintf(
                 'Invalid service manager configuration class provided; received "%s", expected an instance of %s',
-                is_object($config) ? get_class($config) : (is_scalar($config) ? $config : gettype($config)),
+                is_object($config) ? $config::class : (is_scalar($config) ? $config : gettype($config)),
                 ServiceConfigInterface::class
             ));
         }
@@ -258,7 +257,7 @@ class ServiceListener implements ServiceListenerInterface
         if (! $config instanceof ServiceConfig) {
             throw new Exception\RuntimeException(sprintf(
                 'Invalid service manager configuration class provided; received "%s", expected an instance of %s',
-                is_object($config) ? get_class($config) : (is_scalar($config) ? $config : gettype($config)),
+                is_object($config) ? $config::class : (is_scalar($config) ? $config : gettype($config)),
                 ServiceConfig::class
             ));
         }
